[frontend] refactor: questions listing view
