--- frontend/src/utils/questions.ts
@@ -1,2 +1,0 @@
-// temporary so that app can compile, to remove once file is not empty
-export {};
